refactor: 절대경로 수정
