+++ src/pages/@owner/food-truck-form/@section/time-section/ActiveDate.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { useTime } from '@pages/@owner/food-truck-form/hooks/use-time';
 import BottomSheet from '@ui/bottom-sheet/BottomSheet';
 import Calendar from '@components/calendar/Calendar';
-import type { SelectedDate } from '@types/calendar-types';
+import type { SelectedDate } from '@type/calendar-types';
 import FormLayout from '@pages/@owner/food-truck-form/components/FormLayout';
 import ButtonText from '@ui/button-text/ButtonText';
 import ButtonDate from '@ui/button-date/ButtonDate';
