+++ src/pages/filter/hooks/use-filter-logic.ts
@@ -3,7 +3,7 @@ import { useNavigate } from 'react-router-dom';
 import { useAtomValue, useSetAtom } from 'jotai';
 import { isEqual } from 'lodash';
 
-import type { SelectedDate } from '@types/calendar-types';
+import type { SelectedDate } from '@type/calendar-types';
 import {
   filtersAtom,
   initialFilter,
@@ -15,7 +15,7 @@ import type {
   FoodTruckCategoryValue,
   NeedElectricityValue,
   PaymentMethodValue,
-} from '@types/category-types';
+} from '@type/category-types';
 import { ROUTES } from '@router/constant/routes';
 
 export default function useFilterLogic() {
