+++ src/shared/components/calendar/components/CalendarSelectedCircle.tsx
@@ -1,4 +1,4 @@
-import type { CalendarDate } from '@types/calendar-types';
+import type { CalendarDate } from '@type/calendar-types';
 
 import { isSelectedDate } from '@components/calendar/utils/calendar-utils';
