+++ src/shared/components/calendar/hooks/use-calendar.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import type { CalendarDate, SelectedDate } from '@types/calendar-types';
+import type { CalendarDate, SelectedDate } from '@type/calendar-types';
 import { formatCalendarDayToDate } from '@utils/date-formatter';
 
 import {
