+++ src/shared/components/layout/bottom-sheet/BottomSheet.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from '@storybook/react-vite';
-import BottomSheet from '@ui/bottom-sheet/BottomSheet';
+import BottomSheet from '@layout/bottom-sheet/BottomSheet';
 
 const meta: Meta<typeof BottomSheet> = {
-  title: 'Components/BottomSheet',
+  title: 'Components/Layout/BottomSheet',
   component: BottomSheet,
   parameters: {
     layout: 'fullscreen',
