+++ src/shared/store/filter-store.ts
@@ -1,13 +1,13 @@
 import { atom } from 'jotai';
 import { isEqual } from 'lodash';
 
-import type { SelectedDate } from '@types/calendar-types';
+import type { SelectedDate } from '@type/calendar-types';
 import type {
   AvailableQuantityValue,
   FoodTruckCategoryValue,
   NeedElectricityValue,
   PaymentMethodValue,
-} from '@types/category-types';
+} from '@type/category-types';
 
 export interface FilterState {
   schedules: SelectedDate[];
